Fix asyncio memory leak in cancelled tasks and add validation tests
Clear callback chains and exception references in Task._step() to prevent reference retention

Filter cancelled Task objects from event loop's ready queue in BaseEventLoop._run_once()

Add TestTaskMemoryLeak test case to verify proper garbage collection of cancelled tasks

Include module-level docstring explaining test purpose and methodology

Addresses memory management issues reported in Issue #129204
