+++ Lib/asyncio/tasks.py
@@ -263,28 +263,6 @@ class Task(futures._PyFuture):  # Inherit Python Task implementation
                 else:
                     _register_task(self)
 
-    def __step(self, exc=None):
-        if self.done():
-            raise exceptions.InvalidStateError(
-                f'__step(): already done: {self!r}, {exc!r}')
-        if self._must_cancel:
-            if not isinstance(exc, exceptions.CancelledError):
-                exc = self._make_cancelled_error()
-            self._must_cancel = False
-        self._fut_waiter = None
-
-        _enter_task(self._loop, self)
-        try:
-            self.__step_run_and_handle_result(exc)
-        finally:
-            if self.done():
-                # Clear the callback chain and residual references
-                self._callbacks.clear()  # Clean up the callback list
-                self._exception = None  # Release the reference to the exception object
-                if hasattr(self, '_context'):
-                    self._context = None  # Python 3.11+ Clean up context variables
-            _leave_task(self._loop, self)
-            self = None  # Needed to break cycles when an exception occurs.
 
     def __step_run_and_handle_result(self, exc):
         coro = self._coro
